refactor: conversion rates module
