+++ packages/cryptoplease/lib/core/amount.dart
@@ -1,7 +1,6 @@
 import 'package:decimal/decimal.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import 'conversion_rates/bl/repository.dart';
 import 'currency.dart';
 import 'tokens/token.dart';
 
@@ -110,41 +109,11 @@ extension AmountExt on Amount {
 extension CryptoAmountExt on CryptoAmount {
   Token get token => currency.token;
 
-  FiatAmount? toFiatAmount(
-    FiatCurrency currency, {
-    required ConversionRatesRepository ratesRepository,
-  }) {
-    final rate = ratesRepository.readRate(this.currency, to: currency);
-
-    if (rate == null) return null;
-
-    return convert(rate: rate, to: currency) as FiatAmount;
-  }
-
   CryptoAmount copyWithDecimal(Decimal decimal) =>
       copyWith(value: currency.decimalToInt(decimal));
 }
 
 extension FiatAmountExt on FiatAmount {
-  CryptoAmount? toTokenAmount(
-    Token token, {
-    required ConversionRatesRepository ratesRepository,
-  }) {
-    final rate = ratesRepository.readRate(
-      CryptoCurrency(token: token),
-      to: currency,
-    );
-
-    if (rate == null) return null;
-
-    final inverted = rate.inverse.toDecimal(
-      scaleOnInfinitePrecision: token.decimals,
-    );
-
-    return convert(rate: inverted, to: CryptoCurrency(token: token))
-        as CryptoAmount;
-  }
-
   FiatAmount copyWithDecimal(Decimal decimal) =>
       copyWith(value: currency.decimalToInt(decimal));
 }
