+++ packages/cryptoplease/lib/core/conversion_rates/bl/conversion_rates_bloc.dart
@@ -13,8 +13,6 @@ import '../../tokens/token.dart';
 import 'repository.dart';
 
 part 'conversion_rates_bloc.freezed.dart';
-part 'conversion_rates_event.dart';
-part 'conversion_rates_state.dart';
 
 final _logger = Logger('ConversionRatesBloc');
 
@@ -56,3 +54,20 @@ class ConversionRatesBloc extends Bloc<_Event, _State> {
 }
 
 class ConversionRatesRequestException implements Exception {}
+
+@freezed
+class ConversionRatesEvent with _$ConversionRatesEvent {
+  const factory ConversionRatesEvent.refreshRequested({
+    required FiatCurrency currency,
+    required Iterable<Token> tokens,
+  }) = RefreshRequested;
+}
+
+@freezed
+class ConversionRatesState
+    with _$ConversionRatesState
+    implements StateWithProcessingState {
+  const factory ConversionRatesState({
+    @Default(ProcessingStateNone()) ProcessingState processingState,
+  }) = _ConversionRatesState;
+}
