--- packages/cryptoplease/lib/core/conversion_rates/bl/conversion_rates_event.dart
@@ -1,9 +1,0 @@
-part of 'conversion_rates_bloc.dart';
-
-@freezed
-class ConversionRatesEvent with _$ConversionRatesEvent {
-  const factory ConversionRatesEvent.refreshRequested({
-    required FiatCurrency currency,
-    required Iterable<Token> tokens,
-  }) = RefreshRequested;
-}
