--- packages/cryptoplease/lib/core/conversion_rates/bl/conversion_rates_state.dart
@@ -1,10 +1,0 @@
-part of 'conversion_rates_bloc.dart';
-
-@freezed
-class ConversionRatesState
-    with _$ConversionRatesState
-    implements StateWithProcessingState {
-  const factory ConversionRatesState({
-    @Default(ProcessingStateNone()) ProcessingState processingState,
-  }) = _ConversionRatesState;
-}
