+++ packages/cryptoplease/lib/features/outgoing_split_key_payments/src/presentation/oskp_confirmation_screen.dart
@@ -3,7 +3,7 @@ import 'package:dfunc/dfunc.dart';
 import 'package:flutter/material.dart';
 
 import '../../../../core/amount.dart';
-import '../../../../core/conversion_rates/presentation/conversion_rates.dart';
+import '../../../../core/conversion_rates/context_ext.dart';
 import '../../../../core/currency.dart';
 import '../../../../core/presentation/format_amount.dart';
 import '../../../../l10n/device_locale.dart';
